+++ magda-authorization-api/src/Database.ts
@@ -109,7 +109,7 @@ export default class Database {
     async getUserPermissions(id: string): Promise<Permission[]> {
         const result = await this.pool.query(
             `SELECT DISTINCT ON (p.id, op.id)
-                p.id, p.name, p.resource_id, res.uri AS resource_uri, 
+                p.id, p.name, p.resource_id, res.uri AS resource_uri,
                 p.user_ownership_constraint,
                 p.org_unit_ownership_constraint,
                 p.pre_authorised_constraint,